test: unit test of ParseAuthorizationBearerHeader
