+++ api/parse_authorization_bearer_header_test.go
@@ -1,0 +1,33 @@
+package api
+
+import (
+	"marble/marble-backend/models"
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestParseAuthorizationBearerHeader_Norminal(t *testing.T) {
+	header := http.Header{}
+	header.Add("Authorization", "Bearer TOKEN")
+
+	authorization, err := ParseAuthorizationBearerHeader(header)
+	assert.NoError(t, err)
+	assert.Equal(t, authorization, "TOKEN")
+}
+
+func TestParseAuthorizationBearerHeader_EmptyHeader(t *testing.T) {
+
+	authorization, err := ParseAuthorizationBearerHeader(http.Header{})
+	assert.NoError(t, err)
+	assert.Empty(t, authorization)
+}
+
+func TestParseAuthorizationBearerHeader_BadBearerFormat(t *testing.T) {
+	header := http.Header{}
+	header.Add("Authorization", "MalformedBearer")
+
+	_, err := ParseAuthorizationBearerHeader(header)
+	assert.ErrorIs(t, err, models.UnAuthorizedError)
+}
